+++ src/lib.rs
@@ -1,6 +1,7 @@
-#![cfg_attr(feature = "build-docs", feature(doc_auto_cfg))]
-#![cfg_attr(feature = "build-docs", feature(doc_cfg))]
+#![cfg_attr(docsrs, feature(doc_auto_cfg))]
+#![cfg_attr(docsrs, feature(doc_cfg))]
 #![cfg_attr(feature = "no_std", no_std)]
+#![cfg_attr(feature = "unstable", feature(unsize))]
 #![doc = include_str!("../README.md")]
 
 #[cfg(all(
@@ -19,20 +20,19 @@ extern crate alloc;
 
 #[doc(hidden)]
 pub mod arch;
-#[doc(hidden)]
-pub mod thunk;
 
 pub mod bare_closure;
 pub mod cc;
 pub mod jit_alloc;
+pub mod traits;
 
 /// Common imports required to use `closure-ffi`.
 pub mod prelude {
-    #[cfg(feature = "proc_macros")]
     #[doc(inline)]
-    pub use super::bare_closure::bare_dyn;
-    #[doc(inline)]
-    pub use super::bare_closure::{BareFn, BareFnMut, BareFnOnce};
+    pub use super::bare_closure::{
+        BareFn, BareFnAny, BareFnMut, BareFnMutAny, BareFnMutSend, BareFnOnce, BareFnOnceAny,
+        BareFnOnceSend, BareFnSend,
+    };
     #[doc(inline)]
     pub use super::cc;
     #[doc(inline)]